Setup application
Setting up a spring boot webapp with a simple controller that returns
some static authentication data as json.

Later on, we will extend it with spring security authentication to fill
the controller with life.

We create a maven module structure with a parent project, so we can add
libraries containing common authentication classes to make Partner.Net
authentication faster to use.