Implement SAML 2.0 client
The Partner.Net IDP also supports SAML 2.0. This commit adds support for
SAML 2.0 on top of spring security saml2 to the pnet idp client.

The configuration for the saml client is built similar to the OIDC
configuration. The @EnablePartnerNetSaml2 annotation enables the beans
needed for the saml client, and the PartnerNetSaml2Configurer is used
in the spring security config, to configure the saml client.

The configuration adds converters and authentication providers to parse
the Partner.Net response.

Unfortunately spring security does not implement all response validations
like described in the SAML 2.0 specification. So we had to override the
whole AuthenticationProvider to get better validation support.
As the AuthenticationProvider has no access to the HttpRequest, and we
need some stuff from the request, we also had to override the
AuthenticationConverter to get the request in the authentication object.