Implement forceAuthn flag
SAML 2.0 defines the forceAuthn flag, that can be used to force a new
authentication of the user at the IDP.

The application can request this flag before starting the authentication
via the Saml2Utils. Afterwards the
PartnerNetSaml2AuthenticationRequestContextConverter will use this information
and forces authentication.

Requesting of the flag is aligned to be similar to the OIDC mfa request.