Add support for authentication context classes
SAML 2.0 specifies a way to communicate what authentication context class
was used by the IDP to authenticate the user. And the relying party has
the option, to request specific authenticaiton context classes.

We can now use this, to request strong multifactor authentication for our
users if needed. As SAML 2.0 defines a lot of classes, and not all of them
are in use by the IDP, we use the same nist levels as used in the OIDC
client.
Now the application can simply request, strong, or not strong authentication,
and the client chooses the right context classes in the background.