Add url sanitizing on auth success
It can happen, that applications store some saml Processing related
parameters in the saved request. Normally we don't want to have this
parameters in the final url, we redirect the user to after authentication.

So the default implementation simply removes this parameters from the URL.
Applications can override the success handler, when this behaviour is not
appropriate.