Add workaround for serialization problems
In spring-security 5.6.x a problem was introduced, where some classes
stored in the session are not serializable. This can cause problems when
using spring-session to serialize the session into a database.

To overcome this issue, we add a option to register a serialization fix.
This is a Saml2AuthenticationRequestRepository that extracts all the
data from the classes and puts the information in serializable classes
that are stored in the session.

The flag to enable the workaround is available in the EnablePartnerNetSaml2
annotation. A ImportSelector then decides if the workaround repository
needs to be registered or not.

A session serialization mechanism is added to theShowcase application,
to check if serialization of session attributes works.