Replace deprecated AuthenticationRequestFactory
The factory is deprecated and is replaced by a AuthnRequestResolver.
It is much easier to customize the authnrequest using the resolver.
No need to build a context, that stores the data anymore. Only to read
the data later on in another class.
The resolver lets us use a customizer, that can change certain
properties of the authnrequest. So we can keep all the code in
a single class.