Refactor applying of the Saml2 Configurer
Previsouly the configurer was applyed by generating a new instance
and calling the HttpSecurity.apply method manually in the application.
The Configurer then made authorizeRequest calls in the init method.
As this init method is not immediately called when applying the configurer
this could leat to problems where the authorizations are applied after
the authorizations of the application. Meaning we can be applied
after a anyRequest() call, which is illegal.

To overcome this, we add some static method that create a new
instance and apply it directly to the HttpSecurity. The static methods
also add the request authorizations. So they are applied synchronous,
and application developers can decide when to apply them.

Furthermore we add a flag to decide if we want to use authorizeRequest
or authorizeHttpRequest.