Add customizer for OIDC and SAML2
The customizers allow users of the libraries to have full control over
the SAML2 login or OAuth2 Login configurers of Spring Security.
This makes it easier for users to customize the configuration, without
implementing every single configuration option on our own.

Configurations important for Partner.Net specific stuff are implemented
directly in the Partner.Net configurers. Everything else can be directly
applied by the users of the library.