Upgrade to spring boot 3
While upgrading to spring boot 3, we also had to perform the jakarta
namespace migration.
Unfortunately the opensaml library is not compatible with the jakarta
namespace right now. So we had to copy the Endpoint Destination validator
into our own code. Hopefully a future opensaml version can fix this and we
can use the opensaml code again.