Also wrap session on error dispatch
Without the error dispatch, we do not wrap the session if an error
happens. The objects are already serialized in the session, but will
not be deserialized anymore, because of the missing wrapper.
This causes all sort of ClassCastExceptions.

Now that spring security is also configured to be run on ERROR Dispatch,
it is important to also wrap the request here.