Optimize error handling
We also need to provide a log for non login errors. Otherwise it is hard
to find what causes the problem, when a non login error happens.

In addition to that, the failureUrl for the OIDC login is set to the
same URL as for the SAML authentication. It makes no sense to
handle the same sort of error via different URLs.