Add workaround for HttpSessionRequestCache Bug
When the matchingRequestParameterName is set in the request cache.
A optimization is performed that prevents accessing the session on every
request.
Unfortunately this introduces a bug, where the saved request is not used
anymore, when we redirect to it.
When a fix for https://github.com/spring-projects/spring-security/issues/12665
is available, we can remove the workaround.