Add support for max_age to OIDC client
The max_age parameter makes is possible to define the maximum session age
at the IDP. It forces the user to authenticate again, even if remember me
is enabled.
By setting the max_age parameter to 0, it will in fact force a reauthentication
of the user, regardless of when he authenticated.

This is a common use case, whenever you want the user to be redirected to
the login page. For example when the user wants to start a dealerchange,
or authenticate in a different country.

To show how this can work, the showcase application is extended and the
ForceAuthenticationFilter is implemented. That covers exactly this usecase.