Add tenant option to OpenId and SAML2 authentication
Just like the "max-age" parameter, it is now possible to pass a
"tenant" parameter to the authentication mechanismns. If the tenant
does not match the selected tenant, the user has to perform a login.
If the tenant does match the tenant of the IDP session or the
remember-me session, the login is omitted.

The showcase supports this behavior.

(cherry picked from commit 8833cead197023904e47ff273d07bb4b4e7ba505)