Upgrade to Spring Boot 3.3.2
Updated to Spring Boot 3.3.2 in order to reproduce a bug.
The problem is caused by directly instantiating the
DefaultSaml2CredentialsManager that contains annotated
methods and post processing the instantiated manager.

Currently, it is not possible to instantiate the manager
in this fashion and the postProcess call was removed.
As a result, it is now necessary to pass a bean as
Saml2CredentialsManager.

It seems to be caused by the same change, that led to
https://github.com/spring-projects/spring-framework/issues/33286
and was fixed just hours ago.

It's a breaking change! The two credential methods in the
PartnerNetSaml2Configurer are package-private now.
Replace the call by simply passing a
DefaultSaml2CredentialsManager bean initialized with
the configSupplier instead of using the credentials
methods with the configSupplier.

The methods will be restored with the Upgrade to 3.3.3,
if the change in Spring Boot is fixing for our error,
too.